テスト共通化: COM apartment フィクスチャを test_helpers.h に集約
- ComApartmentTest 基底フィクスチャを追加し、CoInitializeEx の HRESULT 追跡 +
  RPC_E_CHANGED_MODE ガードを 1 箇所に集約
- 4 つのテストフィクスチャ (MeasurerParityTest / LayoutTest /
  ImageLoaderTest / ImageLoaderDpiTest) の SetUpTestSuite/TearDownTestSuite
  を削除し ComApartmentTest を継承するよう変更。これにより PR#132 で
  MeasurerParityTest 向けに導入したカウント不整合対策が他 3 フィクスチャ
  にも適用される
- WriteTempFile のコメントから ADD_FAILURE 自明部分を削除
